fix some test

+++ src/main/java/com/project/langrecognizer/service/impl/InMemoryLanguageService.java
@@ -177,6 +177,9 @@ public class InMemoryLanguageService implements LanguageService {
     @LoggingAnnotation
     public String deleteLanguage(final Long id)
             throws ResourceNotFoundException {
+        if(!languageRepository.existsById(id)) {
+            throw new ResourceNotFoundException(NO_LANGUAGE_EXIST_WITH_ID + id);
+        }
         languageRepository.deleteById(id);
         cache.deleteCachedById(id);
         return "language removed!! " + id;
